refactor: TR-2.2: Context Assembly: build prompt context block
